refactor: Service, Controller 분리
