+++ src/main/java/findme/dangdangcrew/meeting/controller/MeetingController.java
@@ -2,7 +2,9 @@ package findme.dangdangcrew.meeting.controller;
 
 import findme.dangdangcrew.global.dto.ResponseDto;
 import findme.dangdangcrew.meeting.dto.*;
+import findme.dangdangcrew.meeting.service.MeetingLeaderService;
 import findme.dangdangcrew.meeting.service.MeetingService;
+import findme.dangdangcrew.meeting.service.MeetingUserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -18,6 +20,7 @@ import java.util.List;
 public class MeetingController {
 
     private final MeetingService meetingService;
+    private final MeetingUserService meetingUserService;
 
     @PostMapping
     @Operation(summary = "모임 생성", description = "모임을 생성합니다.")
@@ -36,30 +39,15 @@ public class MeetingController {
     @PostMapping("/{meetingId}/applications/apply")
     @Operation(summary = "모임 참가 신청", description = "유저가 모임 참가 신청합니다.")
     public ResponseEntity<ResponseDto<MeetingUserResponseDto>> applyMeeting(@PathVariable("meetingId") Long meetingId){
-        MeetingUserResponseDto response = meetingService.applyMeetingByMeetingId(meetingId);
+        MeetingUserResponseDto response = meetingUserService.applyMeetingByMeetingId(meetingId);
         return ResponseEntity.ok(ResponseDto.of(response, "모임 참가 신청을 성공적으로 완료하였습니다."));
     }
 
     @PatchMapping("/{meetingId}/applications/cancel")
     @Operation(summary = "모임 참가 취소", description = "유저가 모임 참가를 취소합니다.")
     public ResponseEntity<ResponseDto<MeetingUserResponseDto>> cancelMeetingApplication(@PathVariable("meetingId") Long meetingId){
-        MeetingUserResponseDto response = meetingService.cancelMeetingApplication(meetingId);
+        MeetingUserResponseDto response = meetingUserService.cancelMeetingApplication(meetingId);
         return ResponseEntity.ok(ResponseDto.of(response, "모임 참가를 성공적으로 취소하였습니다."));
-    }
-
-    @PutMapping("/{meetingId}/applications/status")
-    @Operation(summary = "(모임 생성자) 모임 상태 변경", description = "모임 생성자가 모임 신청 유저들을 확정/취소합니다.")
-    public ResponseEntity<ResponseDto<MeetingUserResponseDto>> updateApplicationStatusByLeader(@PathVariable("meetingId") Long meetingId,
-                                                                               @RequestBody MeetingUserStatusUpdateRequestDto dto){
-        MeetingUserResponseDto response = meetingService.changeMeetingApplicationStatusByLeader(meetingId, dto);
-        return ResponseEntity.ok(ResponseDto.of(response, "모임 신청을 한 유저들의 상태를 성공적으로 변경하였습니다."));
-    }
-
-    @GetMapping("/{meetingId}/applications")
-    @Operation(summary = "(모임 생성자) 모임 신청자 조회", description = "모임 생성자가 모임 참가를 신청한 모든 유저 목록을 조회합니다.")
-    public ResponseEntity<ResponseDto<List<MeetingUserResponseDto>>> getAllMeetingApplications(@PathVariable("meetingId") Long meetingId){
-        List<MeetingUserResponseDto> response = meetingService.readAllApplications(meetingId);
-        return ResponseEntity.ok(ResponseDto.of(response, "모임 참가자 목록을 성공적으로 조회하였습니다.."));
     }
 
     @GetMapping("/{meetingId}/confirmed")
@@ -81,33 +69,5 @@ public class MeetingController {
     public ResponseEntity<ResponseDto<List<MeetingBasicResponseDto>>> getMeetingsByUser(){
         List<MeetingBasicResponseDto> response = meetingService.findMyMeetings();
         return ResponseEntity.ok(ResponseDto.of(response, "가입한 모든 모임을 성공적으로 조회하였습니다."));
-    }
-
-    @PutMapping("/{meetingId}")
-    @Operation(summary = "(모임 생성자) 모임 수정", description = "모임 생성자가 모임을 수정합니다.")
-    public ResponseEntity<ResponseDto<MeetingDetailResponseDto>> updateMeeting(@PathVariable("meetingId") Long meetingId, @RequestBody MeetingRequestDto dto){
-        MeetingDetailResponseDto response = meetingService.updateMeeting(meetingId, dto);
-        return ResponseEntity.ok(ResponseDto.of(response, "모임 수정을 성공적으로 완료하였습니다."));
-    }
-
-    @PatchMapping("/{meetingId}/cancel")
-    @Operation(summary = "(모임 생성자) 모임 삭제", description = "모임 생성자가 모임을 삭제합니다.")
-    public ResponseEntity<ResponseDto> deleteMeeting(@PathVariable("meetingId") Long meetingId){
-        meetingService.deleteMeeting(meetingId);
-        return ResponseEntity.ok(ResponseDto.of(null, "모임 삭제를 성공적으로 완료하였습니다."));
-    }
-
-    @PatchMapping("/{meetingId}/quit")
-    @Operation(summary = "(모임 생성자) 모임 종료", description = "모임 생성자가 모임 종료시 종료 버튼을 누릅니다.")
-    public ResponseEntity<ResponseDto<List<MeetingUserResponseDto>>> quitMeeting(@PathVariable("meetingId") Long meetingId){
-        List<MeetingUserResponseDto> response = meetingService.quitMeeting(meetingId);
-        return ResponseEntity.ok(ResponseDto.of(response, "모임 종료를 성공적으로 완료하였습니다."));
-    }
-
-    @PatchMapping("/{meetingId}/check")
-    @Operation(summary = "(모임 생성자) 모임 참석 여부 변경", description = "모임 생성자가 모임 종료시 모임 참석 여부를 확인합니다.")
-    public ResponseEntity<ResponseDto<List<MeetingUserResponseDto>>> checkUserMeetingStatus(@PathVariable("meetingId") Long meetingId, @RequestBody List<MeetingUserStatusUpdateRequestDto> dto){
-        List<MeetingUserResponseDto> response = meetingService.checkAttendedOrAbsent(meetingId, dto);
-        return ResponseEntity.ok(ResponseDto.of(response, dto.size() + "개의 유저들의 상태를 성공적으로 변경하였습니다."));
     }
 }
